move computed file serializer to own file
